requirements.txt
Django==4.2
pytest
pytest-django
pytest-cov
djangorestframework
django-filter
factory-boy

